# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/user/login": {
      "post": {
        "operationId": "UserController_login",
        "summary": "Login account with username password",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/api/user/all-student-list": {
      "get": {
        "operationId": "UserController_getAllStudentList",
        "summary": "Get all student list",
        "parameters": [
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "default": "createdDate",
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "default": "DESC",
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "classRoomId",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "schoolId",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/api/user/student-list": {
      "get": {
        "operationId": "UserController_getStudentList",
        "summary": "Get student list",
        "parameters": [
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "default": "createdDate",
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "default": "DESC",
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "classRoomId",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "schoolId",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/api/user/teacher-list": {
      "get": {
        "operationId": "UserController_getTeacherList",
        "summary": "Get teacher list",
        "parameters": [
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "default": "createdDate",
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "default": "DESC",
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "classRoomId",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "schoolId",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/api/user/school-list": {
      "get": {
        "operationId": "UserController_getSchoolList",
        "summary": "Get school list",
        "parameters": [
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "default": "createdDate",
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "default": "DESC",
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/api/user/class-list": {
      "get": {
        "operationId": "UserController_getClassList",
        "summary": "Get class list",
        "parameters": [
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "default": "createdDate",
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "default": "DESC",
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/api/user/search": {
      "get": {
        "operationId": "UserController_search",
        "summary": "Get data for author list",
        "parameters": [
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "default": "createdDate",
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "default": "DESC",
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "enum": [
                "APPROVED",
                "PENDING",
                "REJECTED"
              ],
              "type": "string"
            }
          },
          {
            "name": "roleCode",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "default": "USER",
              "enum": [
                "TEACHER",
                "VOLUNTEER",
                "USER",
                "ADMIN",
                "ADMIN_CODE_SERVICE"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/api/user/statistics/{id}": {
      "get": {
        "operationId": "UserController_getStatisticsById",
        "summary": "Get user statistics",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStatistic"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/api/user/search-statistics": {
      "get": {
        "operationId": "UserController_searchStatistics",
        "summary": "Get user statistics",
        "parameters": [
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "default": "createdDate",
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "default": "DESC",
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStatistic"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/api/user/{id}": {
      "get": {
        "operationId": "UserController_getProfileById",
        "summary": "Get user info by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      },
      "put": {
        "operationId": "UserController_updateUser",
        "summary": "Update User",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/api/user/register": {
      "post": {
        "operationId": "UserController_register",
        "summary": "register account with username password",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterDto"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
...
  },
  "info": {
    "title": "Bluequant API APIs document",
    "description": "",
    "version": "3.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "required",
            "default": "dev_admin@gmail.com"
          },
          "password": {
            "type": "string",
            "description": "required",
            "default": "123456"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "LoginResponse": {
        "type": "object",
        "properties": {
          "payload": {
            "$ref": "#/components/schemas/JWTPayload"
          },
          "token": {
            "type": "string",
            "description": "required",
            "default": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjEsImNyZWF0ZUF0IjoiMjAyNC0wMy0yMVQxNDo0OToyNC41ODRaIiwiaWF0IjoxNzExMDMyNTY0LCJleHAiOjE3MTE2MzczNjR9.fL37HAKp-s7IShQmDaB1C54ijhZlUQ1fIBtasnELyPE"
          },
          "expiresIn": {
            "type": "number",
            "description": "required",
            "default": 604800
          }
        },
        "required": [
          "payload",
          "token",
          "expiresIn"
        ]
      },
...
      "DefaultResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "description": "required"
          },
          "data": {
            "$ref": "#/components/schemas/ErrorDataResponse"
          }
        },
        "required": [
          "status",
          "data"
        ]
      },
...
    }
  }
}